Fix file names detection for binaries w/o debug symbols.
We have info about all the text segments from /proc/<pid>/maps in
`struct proc_info` so we can unambiguously detect where EIP/RIP points
to for the case when we don't have debug symbols and bfd_handle is not
helpful enough.  Then we can use this info to print correct file names.